# indexeddb: Shrink values in inbound_group_sessions store, improving performance all around