Fixes #3057.

Change the way we store values in Indexed DB for the crypto store, specifically in the `inbound_group_sessions` store. Reduce the size of the stored values and replace arrays of numbers with base 64 strings. The rationale and some implementation ideas are documented in #3057. (TL;DR: small values are faster, base64 is faster.)

This change:

* Introduces a `MaybeEncrypted` type in `matrix_sdk_indexeddb::crypto_store::indexeddb_serializer` that encapsulates the fact that we serialize differently if the session is encrypted with a store cipher. (Previously, we just serialized an array of numbers in both cases. This was actually quite confusing since these arrays of numbers were holding encoded versions of very different values.) This enum is `untagged` in Serde to avoid increasing the size of Indexed DB records.
* Adds `matrix_sdk_store_encryption::EncryptedValueBase64`, which works the same way as `EncryptedValue`, but stores arrays of numbers as base 64.
* Adds a migration to the new schema to `matrix_sdk_indexeddb::crypto_store::migrations`, including adding the `backed_up_to` property, so we can support a fix for #26892 (see this [comment](https://github.com/element-hq/element-web/issues/26892#issuecomment-1906336076)) without doing another migration.

I also wrote a performance test that runs in a web browser. In the checked-in version, it runs for 2K records, but when I manually increased the number on my machine to 200K, I got these results:

```
Inserting 200000 records with v8 schema took 216,745ms.
Inserting 200000 records with v10 schema took 34,875ms.

Counting 200000 records with v8 schema took 11,786ms.
Counting 200000 records with v10 schema took 469ms.
```

So the speedup is impressive.