@poljar 2024-02-08
Oh no: https://github.com/matrix-org/matrix-rust-sdk/issues/718.