@andybalaam 2024-02-08
> Oh no: #718.

@poljar hmm. I think we could move the JSON-specific stuff into the indexeddb crate if we wanted to?