@andybalaam 2024-02-08
We are already encoding to JSON using `to_string` the question is _what_ are we encoding? Previously it was a `Vec<u8>` and now it's a `String`, which contains base64.