@andybalaam 2024-02-08
If I remember correctly, the `Vec<u8>` we have is a pickled version of an `InboundGroupSession`.