@dkasak 2024-02-08
Just to write down the conclusion: after a quick chat with @andybalaam, only the [unencrypted case](https://github.com/matrix-org/matrix-rust-sdk/pull/3073/files#diff-514d0599fe785ee0062f5a4df3f42b32a0c8ace004a8368e62b08321e18b1e56R158) unnecessarily base64-encodes the payload before storing. The [encrypted case](https://github.com/matrix-org/matrix-rust-sdk/pull/3073/files#diff-514d0599fe785ee0062f5a4df3f42b32a0c8ace004a8368e62b08321e18b1e56R155) a couple of lines above does what is expected, so that's why we were talking past each other.